Base version of the extension
In this base version of chrome extension, The functionality of the extension is to scan the page for span elements, highlight them with color green, change cursor pointer and add event listeners to the span elements.
when the user clicks on any span element, an alert shows up saying that "Span Made Clickable!!!". All clicked links background color are changed to purple.